adding mvc patterns
